Move archive_snapshot to agave-snapshots
