+++ snapshots/Cargo.toml
@@ -34,12 +34,15 @@ solana-clock = { workspace = true }
 solana-genesis-config = { workspace = true }
 solana-hash = { workspace = true }
 solana-lattice-hash = { workspace = true }
+solana-measure = { workspace = true }
+solana-metrics = { workspace = true }
 strum = { workspace = true, features = ["derive"] }
+symlink = { workspace = true }
 tar = { workspace = true }
+tempfile = { workspace = true }
 thiserror = { workspace = true }
 zstd = { workspace = true }
 
 [dev-dependencies]
 agave-logger = { workspace = true }
 assert_matches = { workspace = true }
-tempfile = { workspace = true }
